Fix before_end actions with flavor reservations
Before before_end was set in allocation_candidates. This method
operates on a copy of the reservation, so the side effect of
setting before_end doesn't do anything. Instead we set it in
reserve_resource, which is actually creating the DB entry.

Change-Id: I73a48e7170cafaea4402c79a6ebf0627a7b55d8f
(cherry picked from commit f6da731ad66fbf8074c39000276764f8776809a4)